requirements.txt
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
pillow==11.0.0
PyYAML==6.0.2
scipy==1.15.3
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
